made the date_time function declared as a standard one
